misc-coding/ai-research-platform/sub-projects/idea-database/services/pre_processor/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
markdownify==0.11.*
pyyaml==6.*
langdetect==1.0.*
redis==5.* 